app/requirements.txt
requests
pyttsx3
SpeechRecognition
openai-whisper
pyaudio
soundfile
sounddevice
tkinter